[dev] 공공API 접근 메서드 따로 클래스로 분리 (#13)
